Create nmap_parser.py
Created with ChatGPT kind request:
Could you please write for me a python script using PEP-8 to run nmap and parse it's output? The goal is to take the list of IP addresses and ports, scan them and store in the database results information as: ip address, host name, open and closed ports, services running on those ports and detected operating system.